controllerconfig/controllerconfig/requirements.txt
# Getting values from  https://github.com/openstack/requirements/blob/stable/pike/global-requirements.txt
netaddr>=0.7.13,!=0.7.16  # BSD
keyring>=5.5.1  # MIT/PSF
pyudev  # LGPLv2.1+
psycopg2>=2.5  # LGPL/ZPL
six>=1.9.0  # MIT
iso8601>=0.1.11  # MIT
netifaces>=0.10.4 # MIT
pycrypto>=2.6 # Public Domain
oslo.utils>=3.20.0 # Apache-2.0
PyYAML>=3.1.0
